Rename all _route.tsx files to layout.tsx
